requirements.txt
# Python 3.10.12
Django==5.1.1
djangorestframework==3.15.2
mysqlclient==2.2.4
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
django-soft-delete==1.0.13
python-decouple==3.8